+++ src/forecast_zarr/store.py
@@ -139,14 +139,13 @@ class ForecastStore:
             }
             if variable.standard_name:
                 attrs["standard_name"] = variable.standard_name
-            if variable.group == "derived":
-                attrs["source_variables"] = list(variable.source_short_names)
-                attrs["derivation"] = "forecast-zarr-processor deterministic formula"
-            else:
-                attrs["source_grib_parameters"] = list(variable.source_short_names)
+            attrs["source_grib_parameters"] = list(variable.source_short_names)
             if variable.name == "precipitation_amount":
                 attrs["cell_methods"] = "time: sum"
                 attrs["source_interval_selection"] = "latest startStep ending at valid_time"
+            if variable.name == "precipitation_rate":
+                attrs["cell_methods"] = "time: point"
+                attrs["source_time_statistic"] = "instantaneous"
             if variable.name in {
                 "surface_downwelling_shortwave_flux_in_air",
                 "surface_upwelling_shortwave_flux_in_air",